% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci.R
\name{add_ci}
\alias{add_ci}
\title{Add Confidence Intervals for Predictions to Data Frames.}
\usage{
add_ci(tb, fit, alpha = 0.05, names = NULL, ...)
}
\arguments{
\item{tb}{A tibble or Data Frame on which to make predictions.}

\item{fit}{An object of class lm, glm, or lmerMod. Predictions are
made with this object.}

\item{alpha}{A real number between 0 and 1. Controls the confidence
level of the interval estimates.}

\item{names}{NULL or character vector of length two. If
\code{NULL}, confidence bounds will automatically be named by
\code{add_ci}, otherwise, the lower confidence bound will be
named \code{ciNames[1]} and the upper confidence bound will be
named \code{ciNames[2]}.}

\item{...}{Additional arguments}
}
\value{
A tibble, \code{tb}, with predicted values, upper and lower
    confidence bounds attached.
}
\description{
This is a generic function to append confidence intervals for
predictions of a model fit to a data frame. A confidence interval
is generated for the fitted value of each observation in
\code{tb}. These confidence intervals are then appended to
\code{tb} and returned to the user as a tibble.
}
\examples{
# linear regression
# Append a 50\% confidence interval for the expected response to
#cars
fit1 <- lm(dist ~ speed, data = cars)
add_ci(cars, fit1, alpha = 0.5)

# Poisson regression
# Append a 95\% confidence interval for the expected response to
#cars
fit2 <- glm(dist ~ speed, data = cars, family = "poisson")
add_ci(cars, fit2)

}
