% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_one}
\alias{add_one}
\title{Add 1 to a single numeric}
\usage{
add_one(a)
}
\arguments{
\item{a}{A single numeric value.}
}
\description{
`add_one()` adds 1 to a single numeric value.
}
\examples{

add_one(2)

}
