% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pi_lm.R
\name{add_pi.lm}
\alias{add_pi.lm}
\title{Prediction Intervals for Linear Model Predictions.}
\usage{
\method{add_pi}{lm}(tb, fit, alpha = 0.05, names = NULL,
  log_response = FALSE)
}
\arguments{
\item{tb}{A tibble or Data Frame.}

\item{fit}{An object of class lm. Predictions are made with this
object.}

\item{alpha}{A real number between 0 and 1. Controls the confidence
level of the interval estimates.}

\item{names}{NULL or character vector of length two. If
\code{NULL}, prediction bounds will automatically be named by
\code{add_pi}, otherwise, the lower prediction bound will be
named \code{names[1]} and the upper prediction bound will be
named \code{names[2]}.}

\item{log_response}{logical. If TRUE, prediction intervals will be
generated for the prediction made with a log-linear model:
\eqn{\log(Y) = X\beta + \epsilon}. These intervals will be on
the scale of the original response, Y.}
}
\value{
A tibble, \code{tb}, with predicted values, upper and lower
    prediction bounds attached.
}
\description{
This function is one of the methods for \code{add_pi}. This
function is essentially a wrapper for the \code{predict} function
in \code{R}.
}
\examples{
# linear regression
fit1 <- lm(dist ~ speed, data = cars)
add_pi.lm(cars, fit1, alpha = 0.5)

}
