% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pi.R
\name{add_pi}
\alias{add_pi}
\title{Add Prediction Intervals to Data Frames}
\usage{
add_pi(tb, fit, alpha = 0.05, piNames = NULL, ...)
}
\arguments{
\item{tb}{A tibble or Data Frame on which to make predictions.}

\item{fit}{An object of class lm, glm, or lmerMod. Predictions are
made with this object.}

\item{alpha}{A real number between 0 and 1. Controls the confidence
level of the interval estimates.}

\item{piNames}{NULL or character vector of length two. If
\code{NULL}, prediction bounds will automatically be named by
\code{add_pi}, otherwise, the lower prediction bound will be
named \code{piNames[1]} and the upper prediction bound will be
named \code{piNames[2]}.}

\item{...}{Additional arguments}
}
\value{
A tibble, \code{tb}, with predicted values, upper and lower
    prediction bounds attached.
}
\description{
This is a generic function to append prediction intervals to a data
frame. A prediction interval is made for each observation in
\code{tb} with respect to the model \code{fit}. These intervals are
then appended to \code{tb} and returned to the user as a tibble.
}
\examples{
## linear regression
fit1 <- lm(dist ~ speed, data = cars)
add_pi(cars, fit1, alpha = 0.5)
## Poisson regression
fit2 <- glm(dist ~ speed, data = cars, family = "poisson")
add_pi(cars, fit2)

}
