% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probs_glm.R
\name{add_probs.glm}
\alias{add_probs.glm}
\title{Event Probabilities for Generalized Linear Models}
\usage{
\method{add_probs}{glm}(tb, fit, q, name = NULL, comparison = "<",
  nSims = 200, ...)
}
\arguments{
\item{tb}{A tibble or Data Frame on which to append probabilities}

\item{fit}{An object of class \code{glm}. Predictions are
made with this object.}

\item{q}{A double. A quantile of the response distribution.}

\item{name}{NULL or character vector of length one. If \code{NULL},
probabilities will automatically be named by
\code{add_probs()}, otherwise, the probabilities will be named
\code{name} in the returned tibble}

\item{comparison}{A character vector of length one. If
\code{comparison = "<"}, then Pr(Y|X < q) is
calculated. Must be "<" or ">" for linear, log-linear and
linear mixed models. If \code{fit} is a glm, then
\code{comparison} may also be "<=", ">=", or "=".}

\item{nSims}{A positive integer.}
}
\value{
A tibble, \code{tb}, with predicted values and
    probabilities attached.
}
\description{
This is the method \code{add_probs} uses if the model fit is an
object of class \code{glm}. Probabilities are determined through
simulation, using the same method as \code{add_pi.glm}. Currently,
only logistic and poisson regression are supported.
}
\examples{
fit1 <- glm(dist ~ speed, data = cars, family = "poisson")
add_probs.glm(cars, fit1, q = 40, comparison = "<")
add_probs.glm(cars, fit1, q = 40, comparison = ">=")

}
