% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{addr_clean}
\alias{addr_clean}
\title{clean address text}
\usage{
addr_clean(.x)
}
\arguments{
\item{.x}{a vector of address character strings}
}
\value{
a vector of cleaned addresses
}
\description{
remove excess whitespace keep only letters, numbers, and \code{-}
}
\examples{
addr_clean("224 Woolper Ave Cincinnati OH 45220")
addr_clean("22_4 Woolper Ave. Cincinnati OH 45220")
addr_clean("22\\\\4 W\"oolper Ave; Ci!ncinn&*ati OH 45220")
addr_clean(c("224 Woolper Ave Cincinnati OH 45220",
             "22_4 Woolper Ave. Cincinnati OH 45220",
             "3333 Burnet Ave Cincinnati OH 34230"))
}
