% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{addr_fuzzy_match}
\alias{addr_fuzzy_match}
\alias{fuzzy_match_addr_field}
\title{Fuzzy match addr vectors using field-specific string distances}
\usage{
addr_fuzzy_match(x, y, addr_fields = NULL)

fuzzy_match_addr_field(
  x,
  y,
  addr_field = c("number_prefix", "number_digits", "number_suffix",
    "street_predirectional", "street_premodifier", "street_pretype", "street_name",
    "street_posttype", "street_postdirectional", "place_name", "place_state",
    "place_zipcode"),
  osa_max_dist = 0
)
}
\arguments{
\item{x}{addr vector to match}

\item{y}{addr vector to match to}

\item{addr_fields}{a named vector of OSA maximum distances. Defaults are
used for fields that are not supplied; see Details.}

\item{addr_field}{character name of single addr field to match on}

\item{osa_max_dist}{maximum optimized string alignment distance
used as threshold for matching on single addr field}
}
\value{
a list of integer vectors representing the position of the best
matching address(es) in \code{y} for each address in \code{x}
}
\description{
\code{addr_fuzzy_match()} matches two addr vectors using more than one address
field.

\code{fuzzy_match_addr_field()} matches two addr vectors using a single address
field.

Distances between address tags are defined using optimized string alignment;
see \code{fuzzy_match()} and \code{stringdist::stringdist()} for more details.
}
\details{
Defaults for \code{addr_fields}:
\itemize{
\item number_prefix: 0
\item number_digits: 0
\item number_suffix: 0
\item street_predirectional: 0
\item street_premodifier: 0
\item street_pretype: 0
\item street_name: 1
\item street_posttype: 0
\item street_postdirectional: 0
\item place_name: 0
\item place_state: 0
\item place_zipcode: 0
}

When fuzzy matching \code{street_name}, the "phonetic_street_key"
prefilter is automatically used (see \code{?fuzzy_match}).
}
\examples{
x_addr <- as_addr(c("123 Main St.", "333 Burnet Ave", "3333 Foofy Ave"))
y_addr <- as_addr(c("0000 Main Street", "3333 Burnet Avenue", "222 Burnet Ave"))

# no matches with defaults
addr_fuzzy_match(x_addr, y_addr)

# match on osa_max_dist of 2 for the address number
addr_fuzzy_match(x_addr, y_addr, addr_fields = c("number_digits" = 2))

# ignore address number when matching
addr_fuzzy_match(x_addr, y_addr, addr_fields = c("number_digits" = Inf))


fuzzy_match_addr_field(
  as_addr(c("123 Main St.", "3333 Burnet Ave", "3333 Foofy Ave")),
  as_addr(c("0000 Main Street", "0000 Burnet Avenue", "222 Burnet Ave")),
  addr_field = "street_name", osa_max_dist = 1
)

# empty address fields have an OSA distance of zero and always match
fuzzy_match_addr_field(
  as_addr(c("123 Main St.", "3333 Burnet Ave", "3333 Foofy Ave")),
  as_addr(c("0000 Main Street", "0000 Burnet Avenue", "222 Burnet Ave")),
  addr_field = "number_prefix"
)
}
