% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{addr_hash}
\alias{addr_hash}
\title{Address hashing (hashdress)}
\usage{
addr_hash(x)
}
\arguments{
\item{x}{a character vector of address strings}
}
\value{
a character vector of hashdresses
}
\description{
Addresses are standardized (\code{addr_standardize()}) and then hashed using the md5 algorithm
via \code{digest::hash()}
}
\examples{
addr_hash(c(
  "224 Woolper Avenue Apt 2 Cincinnati OH 45220",
  "224 wOoLpEr Avenue Cincinnati OH 45220",
  "224 Woolper Avenue Apt #2 Cincinnati OH 45220",
  "224 Woolper Ave Cincinnati OH 45220",
  "224 Woolper Av. Cincinnati OH 45220"
))
}
