% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_tigris_match.R
\name{addr_match_tigris_street_ranges}
\alias{addr_match_tigris_street_ranges}
\title{match an addr vector to a tigris street range}
\usage{
addr_match_tigris_street_ranges(x, county = "39061", year = "2022")
}
\arguments{
\item{x}{an addr vector to match}

\item{county}{character string of county identifier}

\item{year}{year of tigris product}
}
\value{
a list of tigris street ranges matching the street name and containing the street number in x;
a NULL value indicates that no street name was matched; a street range tibble with zero rows indicates
that although a street was matched, there was no range containing the street number
}
\description{
match an addr vector to a tigris street range
}
\examples{
addr_match_tigris_street_ranges(as_addr(c("224 Woolper Ave", "3333 Burnet Ave",
                                          "33333 Burnet Ave", "609 Walnut St")))
}
