% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_place.R
\name{addr_place}
\alias{addr_place}
\title{Address place class}
\usage{
addr_place(
  name = NA_character_,
  state = NA_character_,
  zipcode = NA_character_,
  map_state = TRUE
)
}
\description{
\code{addr_place()} objects contain fields for "PlaceName", "StateName",
and "ZipCode".
}
\examples{
addr_place(name = "Cincinnati", state = "OH", zipcode = "45220")
}
