% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_street.R
\name{addr_street}
\alias{addr_street}
\title{Address street class}
\usage{
addr_street(
  predirectional = NA_character_,
  premodifier = NA_character_,
  pretype = NA_character_,
  name = NA_character_,
  posttype = NA_character_,
  postdirectional = NA_character_,
  map_posttype = TRUE,
  map_directional = TRUE,
  map_pretype = TRUE
)
}
\description{
\code{addr_street()} objects contain fields for "StreetNamePreDirectional",
"StreetNamePreModifier", "StreetNamePreType", "StreetName",
"StreetNamePostType", and "StreetNamePostDirectional"
}
\examples{
addr_street(name = "Burnet", posttype = "Ave")
}
