% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/tag.R
\name{usaddress_tag}
\alias{usaddress_tag}
\alias{addr_tag}
\title{Return list of lists of address tags to R.}
\usage{
usaddress_tag(input)

addr_tag(x, clean_address_text = TRUE)
}
\arguments{
\item{input}{a character vector of addresses}

\item{x}{a character vector of addresses}

\item{clean_address_text}{logical; use clean_address_text() to clean addresses prior to tagging?}
}
\value{
a list, the same length as x, of named character vectors of address component tags;
each vector contains all space-separated elements of the cleaned address
and are each named based on inferred address labels (see Details)
}
\description{
Return list of lists of address tags to R.

Tag components of an address string
The address components are tagged using a \href{https://github.com/boydjohnson/usaddress-rs}{rust port} of
\href{https://github.com/datamade/usaddress}{usaddress}.
Component names are based upon the \href{https://www.fgdc.gov/standards/projects/address-data}{United States Thoroughfare, Landmark, and Postal Address Data Standard}.
}
\details{
Possible address labels include:
\itemize{
\item \code{AddressNumberPrefix}
\item \code{AddressNumberSuffix}
\item \code{AddressNumber}
\item \code{BuildingName}
\item \code{CornerOf}
\item \code{IntersectionSeparator}
\item \code{LandmarkName}
\item \code{NotAddress}
\item \code{OccupancyIdentifier}
\item \code{OccupancyType}
\item \code{PlaceName}
\item \code{Recipient}
\item \code{StateName}
\item \code{StreetNamePostDirectional}
\item \code{StreetNamePostType}
\item \code{StreetNamePreDirectional}
\item \code{StreetNamePreModifier}
\item \code{StreetNamePreType}
\item \code{StreetName}
\item \code{SubaddressIdentifier}
\item \code{SubaddressType}
\item \code{USPSBoxGroupID}
\item \code{USPSBoxGroupType}
\item \code{USPSBoxID}
\item \code{USPSBoxType}
\item \code{ZipCode}
}

Find more information about the definitions \href{https://www.fgdc.gov/schemas/address/}{here}
}
\examples{
addr_tag(c("290 Ludlow Avenue Apt #2 Cincinnati OH 45220", "3333 Burnet Ave Cincinnati OH 45219"))
}
