% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cagis_addr}
\alias{cagis_addr}
\title{CAGIS Addresses}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 320453 rows and 7 columns.
}
\usage{
cagis_addr
}
\description{
An example tibble created from the CAGIS addresses with a pre-calculated, unique \code{addr} vector column.
The \code{cagis_parcel_id} column is a list of character vectors because one CAGIS address can correspond to multiple
parcel identifiers.
See \code{data-raw/make_cagis_addr.R} for source code to create data, including filtering criteria:
}
\details{
\itemize{
\item use only addresses that have \code{STATUS} of \code{ASSIGNED} or \code{USING} and are not orphaned (\code{ORPHANFLG == "N"})
\item omit addresses with \code{ADDRTYPE}s that are milemarkers (\code{MM}), parks (\code{PAR}), infrastructure projects (\code{PRJ}),
cell towers (\code{CTW}), vacant or commercial lots (\code{LOT}), and other miscellaneous non-residential addresses (\code{MIS}, \code{RR}, \code{TBA})
\item s2 cell is derived from LONGITUDE and LATITUDE fields in CAGIS address database
}
}
\keyword{datasets}
