% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_address_text}
\alias{clean_address_text}
\title{clean address text}
\usage{
clean_address_text(.x)
}
\arguments{
\item{.x}{a vector of address character strings}
}
\value{
a vector of cleaned addresses
}
\description{
remove excess whitespace; keep only letters, numbers, and \code{-}
}
\examples{
clean_address_text(c(
  "3333 Burnet Ave Cincinnati OH 45219",
  "33_33 Burnet Ave. Cincinnati OH 45219",
  "33\\\\33 B\"urnet Ave; Ci!ncinn&*ati OH 45219",
  "3333 Burnet Ave Cincinnati OH 45219",
  "33_33 Burnet Ave. Cincinnati OH 45219"
))
}
