% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{elh_data}
\alias{elh_data}
\title{Example real-world data with line-one-only addresses}
\usage{
elh_data()
}
\value{
a tibble with 100 rows and 9 columns
}
\description{
The Cincinnati Evicition Hotspots data was downloaded from
\href{https://evictionlab.org/uploads/cincinnati_hotspots_media_report.csv}{Eviction Labs}
and contains characteristics of the top 100 buildings that are responsible for about 25\% of
all eviction filings in Cincinnati (from their "current through 8-31-2024" release).
}
\details{
https://evictionlab.org/eviction-tracking/cincinnati-oh/
}
\examples{
elh_data()
}
