% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{expand_post_type}
\alias{expand_post_type}
\title{Expand street name post type}
\usage{
expand_post_type(x)
}
\arguments{
\item{x}{character vector of length one}
}
\value{
a character vector of length one representing the expanded street name post type
}
\description{
Abbreviations of street type (e.g., "Ave", "St") are converted to
expanded versions (e.g., "avenue", "street").
}
\examples{
expand_post_type("ave")
expand_post_type("av")
expand_post_type("av")
expand_post_type("avenue")
expand_post_type("st")
expand_post_type(NULL)
}
