% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_geocode.R
\name{find_best_range}
\alias{find_best_range}
\title{if multiple in range and parity matches, then choose best based on smallest width of ranges
if ties with the same widths, then choose the range with the closest midpoint}
\usage{
find_best_range(street_number, candidate_ranges)
}
\description{
if multiple in range and parity matches, then choose best based on smallest width of ranges
if ties with the same widths, then choose the range with the closest midpoint
}
