% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode_stage}
\alias{geocode_stage}
\title{Classify geocode stage}
\usage{
geocode_stage(x)
}
\arguments{
\item{x}{a data frame returned by \code{\link[=geocode]{geocode()}}}
}
\value{
an ordered factor with levels \code{none}, \code{street_variant}, \code{street},
\code{range_variant}, \code{range}
}
\description{
Classify geocode results into staged outcomes returned by \code{\link[=geocode]{geocode()}}:
no match, street match, or interpolated range match, distinguishing exact
ZIP-code matches from ZIP-code variant matches.
}
