% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_tigris_match.R
\name{get_tiger_street_ranges}
\alias{get_tiger_street_ranges}
\title{Get tigris street range geography files from census.gov}
\usage{
get_tiger_street_ranges(county, year = "2022")
}
\arguments{
\item{county}{character string of county identifier}

\item{year}{year of tigris product}
}
\value{
a list of tibbles, one for each street name, with \code{TLID}, \code{s2_geography}, \code{from}, and \code{to} columns
}
\description{
Downloaded files are cached in \code{tools::R_user_dir("addr", "cache")}.
Street ranges with missing minimum or maximum address numbers are excluded.
}
\examples{
Sys.setenv("R_USER_CACHE_DIR" = tempfile())
get_tiger_street_ranges("39061")[1001:1004]
}
