% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_tigris_match.R
\name{get_tigris_street_ranges}
\alias{get_tigris_street_ranges}
\title{Get tigris street range geography files from census.gov}
\usage{
get_tigris_street_ranges(county, year = "2022")
}
\arguments{
\item{county}{character string of county identifier}

\item{year}{year of tigris product}
}
\value{
a tibble of street names with a nested column of address range geographies with
min number, max number, id, and geography columns
}
\description{
Get tigris street range geography files from census.gov
}
