% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_part_match.R
\name{match_addr_number}
\alias{match_addr_number}
\title{Match addr_number vectors}
\usage{
match_addr_number(x, y, number_fuzzy_dist = 1L)
}
\arguments{
\item{x, y}{addr_number vectors to match}

\item{number_fuzzy_dist}{integer; maximum optimized string alignment
distance between \code{addr_number} strings in \code{x} and \code{y} to consider a
possible match.}
}
\value{
An \code{addr_number} vector, the same length as \code{x}, containing the
selected match in \code{y} for each element of \code{x}. Unmatched elements are
returned as missing \code{addr_number()} values.
}
\description{
A single \code{addr_number} in \code{y} is chosen for each \code{addr_number} in \code{x}.
If exact matches (using \code{as.character}) are not found,
possible matches within \code{number_fuzzy_dist} are searched for in \code{y}.
If multiple matches are present in \code{y}, the selected match has the
lowest absolute numeric difference from \verb{@digits} in \code{x}; ties are broken
by optimized string alignment (OSA) distance and then by lexicographic
order with digits preceding alphabetic characters.

\code{addr_number} objects with missing \verb{@digits}, or with empty strings
for all of \verb{@prefix}, \verb{@digits}, and \verb{@suffix}, are not matched and
returned as missing instead.
}
\examples{
 x <- addr_number(
   prefix = "",
   digits = as.character(c(1, 10, 228, 11, 22, 22, 22, 10, 99897, NA)),
   suffix = ""
 )

y <- addr_number(
  prefix = "",
  digits = as.character(c(12, 11, 10, 22)),
  suffix = ""
)

match_addr_number(x, y)

match_addr_number(x, y, number_fuzzy_dist = 0L)
}
