% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nad_addr}
\alias{nad_addr}
\title{National Address Database for Hamilton County, OH}
\usage{
nad_addr()
}
\value{
Tibble of addresses from the National Address Database with columns for the concatenated address,
the tagged/parsed addr, the unique id, coordinate placement method, parcel identifier, source,
and s2_geography.
}
\description{
The U.S. Department of Transportation partners with address programs from state,
local, and tribal governments to compile their authoritative data into a database.
Find more information here: https://www.transportation.gov/gis/national-address-database
}
\details{
created with \code{inst/make_NAD_addr.R}; address components are pasted together and parsed/tagged with
\code{as_addr()}; duplicated addresses are removed and only the first NAD UUID is retained
}
\examples{
nad_addr()
x <- addr_match(as_addr("224 Woopler Ave Cinti Oh 45220"), nad_addr()$nad_addr, simplify = TRUE)
x_which_nad <- which(nad_addr()$nad_addr == x)
nad_addr()[x_which_nad, ]
}
