% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{nad_flat_catalog}
\alias{nad_flat_catalog}
\title{Inventory counties in the compressed NAD flat source}
\usage{
nad_flat_catalog(path, member)
}
\description{
Use \code{nad_catalog()} instead of this function directly.
}
\keyword{internal}
