% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nad.R
\name{nad_read}
\alias{nad_read}
\alias{nad_download}
\title{National Address Database (NAD)}
\usage{
nad_read(county, state)

nad_download()
}
\arguments{
\item{county}{character, length one; name of county}

\item{state}{character, length one; name of state}
}
\description{
Download the NAD to the R user's data directory for the addr package
and read data into R by county and state.
}
\details{
The U.S. Department of Transportation partners with address programs from
state, local, and tribal governments to compile their authoritative data
into a database. Find more information here:
\url{https://www.transportation.gov/gis/national-address-database}
The NAD is downloaded as the latest release from the transportation.gov
data portal:
\url{https://data.transportation.gov/dataset/National-Address-Database-NAD-File-Geodatabase/yw36-suxr/about_data}
For the original schema, see
\url{https://www.transportation.gov/sites/dot.gov/files/2023-07/NAD_Schema_202304.pdf}

The NAD does not distinguish between empty and missing address components.
When reading into R, all missing address components are replaced with an
empty string (\code{""}) \emph{except} for address number (digits), street name,
and ZIP code.
}
\examples{
\dontrun{
  nad_read("Hamilton", "OH")
}

# example data preloaded for Hamilton County, OH
nad_example_data()
}
