% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr-package.R
\docType{data}
\name{practice_addr_data}
\alias{practice_addr_data}
\title{Fake parameter a and b data.}
\format{
A data frame with 10 rows and 2 variables:
\describe{
\item{param_a}{randomly generated numbers with no units}
\item{param_b}{randomly generated numbers with no units}
}
}
\source{
Me randomly typing on my keyboard
}
\usage{
data(practice_addr_data)
}
\description{
A fabricated dataset containing parameters a and b that were generated at random,
solely for the purpose of illustrating how to create a package with functions and
data, as well as its documentation.
}
\examples{
 data(practice_addr_data)
 param_a <- practice_addr_data$param_a
}
\keyword{datasets}
