% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_usaddress.R
\name{tag_usaddress}
\alias{tag_usaddress}
\title{Tag US addresses}
\usage{
tag_usaddress(x = NA_character_, clean = TRUE)
}
\arguments{
\item{x}{character string of addresses}

\item{clean}{logical; clean address text with clean_address_text()
before tagging?}
}
\value{
a list of vectors of named address tags
}
\description{
Addresses are tagged using the usaddress conditional random field
in a \href{https://github.com/boydjohnson/usaddress-rs}{rust port} of
\href{https://github.com/datamade/usaddress}{usaddress}.
Possible address labels include: \code{AddressNumberPrefix},
\code{AddressNumberSuffix}, \code{AddressNumber}, \code{BuildingName},
\code{CornerOf}, \code{IntersectionSeparator}, \code{LandmarkName},
\code{NotAddress}, \code{OccupancyIdentifier}, \code{OccupancyType},
\code{PlaceName}, \code{Recipient}, \code{StateName}, \code{StreetNamePostDirectional},
\code{StreetNamePostType}, \code{StreetNamePreDirectional}, \code{StreetNamePreModifier},
\code{StreetNamePreType}, \code{StreetName}, \code{SubaddressIdentifier}, \code{SubaddressType}
\code{USPSBoxGroupID}, \code{USPSBoxGroupType}, \code{USPSBoxID}, \code{USPSBoxType},
or \code{ZipCode}.
}
\details{
Find more information about the definitions at
\url{https://www.fgdc.gov/schemas/address/}
}
\examples{
tag_usaddress(c("290 Ludlow Avenue Apt 2 Cincinnati OH 45220",
                "3333 Burnet Ave Cincinnati Ohio 45219",
                NA, ""))
}
