% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_geography.R
\name{tiger_block_groups}
\alias{tiger_block_groups}
\title{get s2_geography for census block groups}
\usage{
tiger_block_groups(state, year)
}
\arguments{
\item{state}{census FIPS state identifier}

\item{year}{vintage of TIGER/Line block group geography files}
}
\value{
a tibble with \code{GEOID} and \code{s2_geography} columns
}
\description{
get s2_geography for census block groups
}
