% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger.R
\name{tiger_download}
\alias{tiger_download}
\title{download tiger files to the R user's data directory for the addr package}
\usage{
tiger_download(x)
}
\arguments{
\item{x}{filename relative to ftp://ftp2.census.gov/geo/tiger/}
}
\description{
download tiger files to the R user's data directory for the addr package
}
\keyword{internal}
