% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger.R
\name{tiger_street_ranges}
\alias{tiger_street_ranges}
\title{Get s2_geography for tiger street ranges}
\usage{
tiger_street_ranges(county, year = "2022")
}
\arguments{
\item{county}{character string of county identifier}

\item{year}{year of tigris product}
}
\value{
a list of tibbles, one for each street name, with \code{TLID}, \code{s2_geography}, \code{from}, and \code{to} columns
}
\description{
Street ranges with missing minimum or maximum address numbers are excluded.
}
\keyword{internal}
