% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{usaddress_tag}
\alias{usaddress_tag}
\title{Run usaddress::parse in Rust on a character vector in R}
\usage{
usaddress_tag(x)
}
\description{
Use \code{tag_usaddress()} instead of this function directly.
}
\keyword{internal}
