% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_part_match.R
\name{zipcode_variant}
\alias{zipcode_variant}
\title{Create ZIP code variants}
\usage{
zipcode_variant(x, variant = c("minus1", "plus1", "sub5", "sub4", "swap"))
}
\arguments{
\item{x}{character length one; five digit ZIP code}

\item{variant}{character one or more variants to create; see description}
}
\value{
character vector of five digit ZIP code variants
}
\description{
An input ZIP code is used to generate variants (for, e.g., 45220):
\itemize{
\item \code{minus1}: substracting one from zipcode (45219)
\item \code{plus1}: adding one to zipcode (45219)
\item \code{sub5}: substituting the fifth digit of the ZIP code (45221, 45222, 45223, 45224, 45225, 45226, 45227, 45228, 45229)
\item \code{sub4}: substituting the fourth digit of the ZIP code (45200, 45210, 45230, 45240, 45250, 45260, 45270, 45280, 45290)
\item \code{swap}: swapping the second and third digits of the ZIP code (42520)
}

More than one variant type can be created at once and variants will be
returned in the same order as they were requested (see examples).
}
\examples{
zipcode_variant("45220")

# order matters!
zipcode_variant("45220", c("minus1", "plus1"))
zipcode_variant("45220", c("plus1", "minus1"))

zipcode_variant("45220", "sub5")

zipcode_variant("45220", "swap")
}
