% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ci.R
\name{add_ci}
\alias{add_ci}
\title{Add CI generic function}
\usage{
add_ci(tb, fit, alpha = 0.05, ciNames = NULL, ...)
}
\arguments{
\item{tb}{A tibble or Data Frame}

\item{fit}{An object or class lm, glm, or lmerMod}

\item{alpha}{A real number between 0 and 1}

\item{ciNames}{NULL or character vector of length two}
}
\value{
A data frame, tb, with predicted values, upper and lower confidence
  bounds
}
\description{
This is a generic function to append confidence intervals for the predicted
values of a model fit to a data frame.
}
