% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr-package.R
\docType{package}
\name{addr-package}
\alias{addr-package}
\title{addr: Adds Two Numbers Together}
\description{
A simple prectice package for GEOG 712 that contains one function named addr that is used to add two numbers together.
}
\author{
\strong{Maintainer}: Claire Estey \email{esteyc@mcmaster.ca} (\href{https://orcid.org/0000}{ORCID})

}
\keyword{internal}
