% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{addr-package}
\alias{addr-package}
\title{addr: Clean, Parse, Harmonize, and Hash Messy Real-World Addresses}
\description{
Addresses that were not validated at the time of collection are often heterogenously formatted, making them difficult to compare or link to other sets of addresses. The addr package is designed to clean character strings of addresses, use the `usaddress` library to tag address components, and paste together select components to create a normalized address. Normalized addresses can be hashed to create hashdresses that can be used to merge with other sets of addresses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cole-brokamp/addr}
  \item Report bugs at \url{https://github.com/cole-brokamp/addr/issues}
}

}
\author{
\strong{Maintainer}: Cole Brokamp \email{cole@colebrokamp.com} (\href{https://orcid.org/0000-0002-0289-3151}{ORCID})

}
