% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr-package.R
\docType{package}
\name{addr-package}
\alias{addr}
\alias{addr-package}
\title{addr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
