% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr.R
\name{addr}
\alias{addr}
\title{A function for adding two numbers}
\usage{
addr(a = 1, b = 2)
}
\arguments{
\item{a}{A real number}

\item{b}{A real number}
}
\value{
the sum of \code{a} and \code{b}
}
\description{
A practice function for GEOG 712 to make sure I know how to write functions.
It simply adds two numbers together.
}
\examples{
addr(4, 5)
addr(9, 17)
}
