% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr.R
\name{addr}
\alias{addr}
\title{Create a new addr object}
\usage{
addr(
  x = character(),
  clean_address_text = TRUE,
  expand_street_type = TRUE,
  clean_zip_code = TRUE
)
}
\arguments{
\item{x}{a character vector of address strings}

\item{clean_address_text}{logical; use \code{clean_address_text()} to clean addresses prior to tagging?}

\item{clean_zip_code}{logical; truncate tagged ZIP Code to 5 characters and set to missing if any of them are not digits?}

\item{expand_street_name_post_type}{logical; use \code{expand_post_type()} to expand \code{StreetNamePostType} tags?}
}
\description{
An addr object is created by converting messy, real-world mailing addresses in a
character vector into standardized address tags for comparison and lookup with \code{addr_tag()}.
By default, address strings are cleaned with \code{addr_clean()},
ZIP codes are restricted to the first five digits and set to missing
if they contain any non-digit characters,
and the street name post types are expanded (e.g., "Ave" -> "Avenue").
}
\details{
In the case of an address having more than one word for a tag (e.g., "Riva Ridge" for \code{StreetName}),
then these are concatenated together, separated by a space in the order they appeared in the address.
}
\examples{
addr(c("3333 Burnet Ave Cincinnati OH 45219", "1324 Burnet Ave Cincinnati OH 45219"))
}
