% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr.R
\name{addr}
\alias{addr}
\title{Create a new addr object}
\usage{
addr(
  x = character(),
  clean_address_text = TRUE,
  expand_street_type = TRUE,
  abbrev_cardinal_dir = TRUE,
  clean_zip_code = TRUE
)
}
\arguments{
\item{x}{a character vector of address strings}

\item{clean_address_text}{logical; use \code{clean_address_text()} to clean address text prior to tagging?}

\item{expand_street_type}{logical; use \code{expand_post_type()} to expand \code{StreetNamePostType} tags?}

\item{abbrev_cardinal_dir}{logical; abbreviate cardinal directions?}

\item{clean_zip_code}{logical; remove any non-digit (or hyphen) characters and truncate tagged ZIP Code to 5 characters?}
}
\description{
An addr object is created by converting messy, real-world mailing addresses in a
character vector into standardized address tags for comparison and lookup with \code{addr_tag()}.
By default, text is cleaned with \code{clean_address_text()},
ZIP codes are restricted to the first five digits and set to missing
if they contain any non-digit characters,
street name post types are expanded (e.g., "Ave" -> "Avenue"),
cardinal directions are abbreviated (e.g. "west" -> "W"),
and non-digit characters are removed from the street number.
}
\details{
In the case of an address having more than one word for a tag (e.g., "Riva Ridge" for \code{StreetName}),
then these are concatenated together, separated by a space in the order they appeared in the address.
}
\examples{
addr(c("3333 Burnet Ave Cincinnati OH 45229", "1324 Burnet Ave Cincinnati OH 45229"))
}
