% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr.R
\name{addr}
\alias{addr}
\title{Address class}
\usage{
addr(number, street, place)
}
\arguments{
\item{number}{a addr_number object}

\item{street}{a addr_street_object}

\item{place}{a addr_street_object}
}
\description{
\code{addr()} combines \code{addr_number()}, \code{addr_street()}, and \code{addr_place()} into a
single address object. The structures for \code{addr()} and the \code{addr_} classes are
derived as a subset of the United States Thoroughfare, Landmark, and Postal
Address Data Standard that is relevant for residential thoroughfare
addresses.
}
\examples{
addr(
  addr_number(digits = "290"),
  addr_street(name = "Burnet", posttype = "Ave"),
  addr_place(name = "Cincinnati", state = "OH", zipcode = "45229")
)
addr(
  addr_number(digits = "200"),
  addr_street(
    predirectional = "west",
    premodifier = "Old",
    pretype = "US",
    name = "50",
    posttype = "avenue",
    postdirectional = "east",
    map_directional = TRUE,
    map_pretype = TRUE,
    map_posttype = TRUE
  ),
  addr_place(name = "Cincinnati", state = "ohio", zipcode = "45220")
)

# addr_* objects are recycled and omitted fields are missing
addr(
  addr_number(digits = c("290", "200", "3333", "111")),
  addr_street(
    name = c("Burnet", "Main", "Ludlow", "State Route 32"),
    posttype = c("Ave", "St", "Ave", NA_character_)
  ),
  addr_place(name = "Cincinnati", state = "OH")
)|>
  str()
}
