% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_part.R, R/addr.R
\name{addr_number}
\alias{addr_number}
\alias{addr_street}
\alias{addr_place}
\alias{addr}
\title{addr classes}
\usage{
addr_number(
  prefix = NA_character_,
  digits = NA_character_,
  suffix = NA_character_
)

addr_street(
  predirectional = NA_character_,
  premodifier = NA_character_,
  pretype = NA_character_,
  name = NA_character_,
  posttype = NA_character_,
  postdirectional = NA_character_,
  map_posttype = TRUE,
  map_directional = TRUE,
  map_pretype = TRUE,
  map_ordinal = TRUE
)

addr_place(
  name = NA_character_,
  state = NA_character_,
  zipcode = NA_character_,
  map_state = TRUE
)

addr(number = addr_number(), street = addr_street(), place = addr_place())
}
\arguments{
\item{prefix}{(often fractional) appears before digits}

\item{digits}{primary street number for the address}

\item{suffix}{(often letter/part) attached after digits}

\item{predirectional}{direction before name}

\item{premodifier}{descriptive modifier before name}

\item{pretype}{type/classification before name}

\item{name}{city, town, or municipality name}

\item{posttype}{type/classification after name}

\item{postdirectional}{direction after name}

\item{map_posttype}{logical; map posttype to abbreviations?}

\item{map_directional}{logical; map pre- and post-directional
to abbreviations?}

\item{map_pretype}{logical; map pretype to abbreviations?}

\item{map_ordinal}{logical; map ordinal street names to abbreviations?}

\item{state}{state or territory abbreviation}

\item{zipcode}{five digit ZIP code}

\item{map_state}{logical; map state to abbreviations?}

\item{number}{an addr_number vector}

\item{street}{an addr_street vector}

\item{place}{an addr_place vector}
}
\value{
An addr, addr_number, addr_street, or addr_place vector
}
\description{
\code{addr()} combines \code{addr_number()}, \code{addr_street()}, and \code{addr_place()} into a
single addr vector. The structures for \code{addr()} and the \code{addr_} classes are
derived as a subset of the United States Thoroughfare, Landmark, and Postal
Address Data Standard that is relevant for residential, numbered thoroughfare
addresses:
\itemize{
\item \code{addr_number()} vectors contain fields for "AddressNumberPrefix",
"AddressNumber", and "AddressNumberSuffix".
\item \code{addr_street()} vectors contain fields for "StreetNamePreDirectional",
"StreetNamePreModifier", "StreetNamePreType", "StreetName",
"StreetNamePostType", and "StreetNamePostDirectional".
\item \code{addr_place()} vectors contain fields for "PlaceName", "StateName",
and "ZipCode".
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{<addr>
 @ number: <addr_number> function ()
 .. @ prefix: chr ""
 .. @ digits: chr ""
 .. @ suffix: chr ""
 @ street: <addr_street> function ()
 .. @ predirectional : chr ""
 .. @ premodifier    : chr ""
 .. @ pretype        : chr ""
 .. @ name           : chr ""
 .. @ posttype       : chr ""
 .. @ postdirectional: chr ""
 @ place : <addr_place> function ()
 .. @ name   : chr ""
 .. @ state  : chr ""
 .. @ zipcode: chr ""
}\if{html}{\out{</div>}}
}
\details{
All field values must must be a character vector of at least length one
(including missing values); length one fields will be
recycled to match the length of other fields.
}
\examples{
# define a new addr_number vector
addr_number(digits = "290")
addr_number(prefix = "N", digits = "290", suffix = "A")

# define a new addr_street vector
addr_street(name = "Burnet", posttype = "Ave")

# street names are automatically mapped to abbreviations
addr_street(predirectional = "North", name = "Fifth", posttype = "Street")

# define a new addr_place vector
addr_place(name = "Cincinnati", state = "OH", zipcode = "45220")

# define a new addr vector
addr(
  addr_number(digits = "290"),
  addr_street(name = "Burnet", posttype = "Ave"),
  addr_place(name = "Cincinnati", state = "OH", zipcode = "45229")
)

# define a more complicated addr vector
# and explicltly specify empty components to avoid NA
addr(
  addr_number(prefix = "", digits = "200", suffix = ""),
  addr_street(
    predirectional = "west",
    premodifier = "Old",
    pretype = "US",
    name = "50",
    posttype = "avenue",
    postdirectional = "east",
    map_directional = TRUE,
    map_pretype = TRUE,
    map_posttype = TRUE
  ),
  addr_place(name = "Cincinnati", state = "ohio", zipcode = "45220")
)

# addr_* vectors are recycled and omitted fields are missing
addr(
  addr_number(digits = c("290", "200", "3333", "111")),
  addr_street(
    name = c("Burnet", "Main", "Ludlow", "State Route 32"),
    posttype = c("Ave", "St", "Ave", NA_character_)
  ),
  addr_place(name = "Cincinnati", state = "OH")
)
}
