% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{addr_clean}
\alias{addr_clean}
\title{clean address text}
\usage{
addr_clean(.x)
}
\arguments{
\item{.x}{a vector of address character strings}
}
\value{
a vector of cleaned addresses
}
\description{
remove excess whitespace keep only letters, numbers, and \code{-}
}
\examples{
addr_clean(c(
    "3333 Burnet Ave Cincinnati OH 45219",
    "33_33 Burnet Ave. Cincinnati OH 45219",
    "33\\\\33 B\"urnet Ave; Ci!ncinn&*ati OH 45219",
    "3333 Burnet Ave Cincinnati OH 45219",
    "33_33 Burnet Ave. Cincinnati OH 45219"
  ))
}
