% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_fuzzy_left_join.R
\name{addr_fuzzy_left_join}
\alias{addr_fuzzy_left_join}
\title{Left join two data frames using fuzzy addr matching}
\usage{
addr_fuzzy_left_join(
  x,
  y,
  by = "addr",
  addr_fields = NULL,
  suffix = c(".x", ".y"),
  progress = interactive()
)
}
\arguments{
\item{x, y}{data frames or tibbles with an addr column}

\item{by}{addr column name in \code{x} (and \code{y} if the same);
or a length-2 character vector of \code{c(x_col, y_col)}}

\item{addr_fields}{a named vector of osa_max_distances; if max distances
for each addr tag field is not present a default will be used (see details).}

\item{suffix}{character vector of length 2 used to suffix duplicate columns}

\item{progress}{logical; show progress bar while processing matched
ZIP groups?}
}
\value{
a data frame with left-join semantics; note that \emph{row order will
be changed compared to x}
}
\description{
This wraps around the addr fuzzy matching helpers and
returns a left-join style result. The addr columns are matched by index and
rows are expanded for one-to-many or many-to-many matches.

See \code{addr_match()} and \code{addr_left_join()} for an alternative strategy
to matching addr and joining data frames with addr columns
that is faster, but is limited to one best match instead of
multi-matches.
}
\details{
addr_fuzzy_left_join works by matching addresses grouped by ZIP codes, so specified
osa_max_distances for any place fields are ignored
Defaults for \code{addr_fields}:
\itemize{
\item number_prefix: 0
\item number_digits: 0
\item number_suffix: 0
\item street_predirectional: 0
\item street_premodifier: 0
\item street_pretype: 0
\item street_name: 1
\item street_posttype: 0
\item street_postdirectional: 0
}
}
\examples{
my_addr <-
  tibble::tibble(address = voter_addresses()[1:10],
                 addr = as_addr(address),
                 id = sprintf("id_\%04d", seq_len(10)))
the_addr <- nad_example_data()
addr_fuzzy_left_join(my_addr, the_addr, c("addr", "nad_addr"))
}
