% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{addr_fuzzy_match}
\alias{addr_fuzzy_match}
\alias{fuzzy_match_addr_field}
\title{Fuzzy match addr objects using field-specific string distances}
\usage{
addr_fuzzy_match(x, y, addr_fields = NULL)

fuzzy_match_addr_field(
  x,
  y,
  addr_field = c("number_prefix", "number_digits", "number_suffix",
    "street_predirectional", "street_premodifier", "street_pretype", "street_name",
    "street_posttype", "street_postdirectional", "place_name", "place_state",
    "place_zipcode"),
  osa_max_dist = 0
)
}
\arguments{
\item{x}{addr object to match}

\item{y}{addr object to match to}

\item{addr_fields}{a named vector of osa_max_distances; if max distances
for each addr tag field is not present a default will be used (see details).}

\item{addr_field}{character name of single addr field to match on}

\item{osa_max_dist}{maximum optimized string alignment distance
used as threshold for matching on single addr field}
}
\value{
a list of integer vectors representing the position of the best
matching address(es) in \code{y} for each address in \code{x}
}
\description{
\code{addr_fuzzy_match()} matches two addr vectors using more than one address tag

\code{fuzzy_match_addr_field()} matches two addr vectors using a single address
tag

Distances between address tags are defined using optimized string alignment;
see \code{fuzzy_match()} and \code{stringdist::stringdist()} for more details.
}
\details{
Defaults for \code{addr_fields}:
\itemize{
\item number_prefix: 0
\item number_digits: 0
\item number_suffix: 0
\item street_predirectional: 0
\item street_premodifier: 0
\item street_pretype: 0
\item street_name: 1
\item street_posttype: 0
\item street_postdirectional: 0
\item place_name: 0
\item place_state: 0
\item place_zipcode: 0
}
}
\examples{
addr_fuzzy_match(
  as_addr(c("123 Main St.", "3333 Burnet Ave", "3333 Foofy Ave")),
  as_addr(c("0000 Main Street", "0000 Burnet Avenue", "222 Burnet Ave")),
  addr_fields = c("number_digits" = Inf, "street_name" = 1)
)

# only match on 'line one' portion of address
addr_fuzzy_match(as_addr(c("3333 Burnet Ave", "3333 Foofy Ave")),
                    as_addr(c("333 Foofy Avenue", "3333 Burnet Avenue", "3333 Burnet Way")),
                    addr_fields = c(
                      place_name = Inf,
                      place_state = Inf,
                      place_zipcode = Inf))

fuzzy_match_addr_field(
  as_addr(c("123 Main St.", "3333 Burnet Ave", "3333 Foofy Ave")),
  as_addr(c("0000 Main Street", "0000 Burnet Avenue", "222 Burnet Ave")),
  addr_field = "street_name", osa_max_dist = 1
)

# empty address fields have an OSA distance of zero and always match
fuzzy_match_addr_field(
  as_addr(c("123 Main St.", "3333 Burnet Ave", "3333 Foofy Ave")),
  as_addr(c("0000 Main Street", "0000 Burnet Avenue", "222 Burnet Ave")),
  addr_field = "number_prefix"
)
}
