% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_hash.R
\name{addr_hash}
\alias{addr_hash}
\title{hash addr vectors}
\usage{
addr_hash(x)
}
\arguments{
\item{x}{an addr vector (or character vector coerced to an addr vector)}
}
\value{
character vector
}
\description{
Addr vectors are converted to character strings and hashed using the md5 algorithm from the digest package
}
\examples{
addr_hash(c("3333 Burnet Ave Cincinnati OH 45229", "1324 Burnet Ave Cincinnati OH 45229"))
addr_hash(c("3333 Burnet Ave Cincinnati OH 45229", "3333 burnet Avenue Cincinnati OH 45229"))
}
