% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_match.R
\name{addr_match}
\alias{addr_match}
\title{matching addr vectors}
\usage{
addr_match(x, ref_addr)
}
\arguments{
\item{x}{an addr vector to match}

\item{ref_addr}{an addr vector to search for matches in}
}
\value{
a list of matching reference addr vectors for each addr in x
}
\description{
For an addr vector, the string distances are calculated between a reference addr vector (\code{ref_addr}).
A list of matching reference addr vectors less than or equal to the specified
\href{https://en.wikipedia.org/wiki/Damerau\%E2\%80\%93Levenshtein_distance#Optimal_string_alignment_distance}{optimal string alignment}
distances are returned.
See \code{stringdist::stringdist-metrics} for more details on string metrics and the optimal string alignment (\code{osa}) method.
}
\examples{
addr(c("3333 Burnet Ave Cincinnati OH 45229", "5130 RAPID RUN RD CINCINNATI OHIO 45238")) |>
  addr_match(cagis_addr()$cagis_addr) |>
  tibble::enframe(name = "input_addr", value = "ca") |>
  dplyr::mutate(ca = purrr::list_c(ca)) |>
  dplyr::left_join(cagis_addr(), by = c("ca" = "cagis_addr"))
}
