% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_match.R
\name{addr_match}
\alias{addr_match}
\title{matching addr vectors}
\usage{
addr_match(x, ref_addr)
}
\arguments{
\item{x}{an addr vector to match}

\item{ref_addr}{an addr vector to search for matches in}
}
\value{
a list of matching reference addr vectors for each addr in x
}
\description{
For an addr vector, the string distances are calculated between a reference addr vector (\code{ref_addr}).
A list of matching reference addr vectors less than or equal to the specified
\href{https://en.wikipedia.org/wiki/Damerau\%E2\%80\%93Levenshtein_distance#Optimal_string_alignment_distance}{optimal string alignment}
distances are returned.
See \code{stringdist::stringdist-metrics} for more details on string metrics and the optimal string alignment (\code{osa}) method.
}
