% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_match.R
\name{addr_match}
\alias{addr_match}
\title{matching addr vectors}
\usage{
addr_match(
  x,
  y,
  max_dist_street_number = 0,
  max_dist_street_name = 1,
  max_dist_street_type = 0,
  simplify = FALSE
)
}
\arguments{
\item{x}{an addr vector to match}

\item{y}{an addr vector to search for matches in}
}
\value{
a named list of possible addr matches for each addr in \code{x};
a list value of NULL means the zip code was not matched and
a list value of a zero-length addr vector means the zip code was matched,
but the street number, name, and type were not matched
}
\description{
Optimized String Alignment (OSA) distances are used to
choose a set of matching reference addr with flexible, address
tag-specific thresholds.
See \code{fuzzy_match()}/\code{fuzzy_match_addr_field()} for more details.
addr vectors are matched in groups by five digit ZIP codes.
}
\examples{
addr(c("3333 Burnet Ave Cincinnati OH 45229",
       "5130 RAPID RUN RD CINCINNATI OHIO 45238")) |>
  addr_match(nad_addr()$nad_addr)
}
