% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_tiger_match.R
\name{addr_match_tiger_street_ranges}
\alias{addr_match_tiger_street_ranges}
\title{Match an addr vector to TIGER street ranges}
\usage{
addr_match_tiger_street_ranges(
  x,
  county = "39061",
  year = "2022",
  street_only_match = c("none", "all", "closest"),
  summarize = c("none", "union", "centroid")
)
}
\arguments{
\item{x}{an addr vector to match}

\item{county}{character string of county identifier}

\item{year}{year of tigris product}

\item{street_only_match}{for addresses that match a TIGER street name, but have street numbers that don't
intersect with ranges of potential street numbers, return \code{"none"}, \code{"all"}, or the \code{"closest"} range geographies}

\item{summarize}{optionally summarize matched street ranges as their union or centroid}
}
\value{
a list of matched tigris street range tibbles;
a NULL value indicates that no street name was matched; if \code{street_only_match} is FALSE,
a street range tibble with zero rows indicates that although a street was matched,
there was no range containing the street number
}
\description{
Match an addr vector to TIGER street ranges
}
\examples{
my_addr <- as_addr(c("224 Woolper Ave", "3333 Burnet Ave", "33333 Burnet Ave", "609 Walnut St"))

addr_match_tiger_street_ranges(my_addr, county = "39061", street_only_match = FALSE)

addr_match_tiger_street_ranges(my_addr, county = "39061", summarize = "centroid")

addr_match_tiger_street_ranges(my_addr, county = "39061", summarize = "centroid") |>
  dplyr::bind_rows() |>
  dplyr::mutate(census_bg_id = s2_join_tiger_bg(s2::as_s2_cell(s2_geography)))
}
