% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_number.R
\name{addr_number}
\alias{addr_number}
\title{Address number class}
\usage{
addr_number(
  prefix = NA_character_,
  digits = NA_character_,
  suffix = NA_character_
)
}
\description{
\code{addr_number()} objects contain fields for "AddressNumberPrefix",
"AddressNumber", and "AddressNumberSuffix".
}
\examples{
addr_number(digits = "290")
addr_number(prefix = "N", digits = "290", suffix = "A")
}
