% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_place.R
\name{addr_place}
\alias{addr_place}
\title{Address place class}
\usage{
addr_place(
  name = character(0),
  state = character(0),
  zipcode = character(0),
  map_state = TRUE
)
}
\description{
Address place class
}
