% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_place.R
\name{addr_place}
\alias{addr_place}
\title{Address place class}
\usage{
addr_place(
  name = NA_character_,
  state = NA_character_,
  zipcode = NA_character_,
  map_state = TRUE
)
}
\arguments{
\item{name}{city, town, or municipality name}

\item{state}{state or territory abbreviation}

\item{map_state}{logical; map state to abbreviations?}
}
\value{
a addr_place object
}
\description{
\code{addr_place()} objects contain fields for "PlaceName", "StateName",
and "ZipCode".
Each must be a character vector of at least length one
(including missing values); length one fields will be
recycled to match the length of other fields.
}
\examples{
addr_place(name = "Cincinnati", state = "OH", zipcode = "45220")
}
