% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{addr_standardize}
\alias{addr_standardize}
\title{Address standardization}
\usage{
addr_standardize(x)
}
\arguments{
\item{x}{a character vector of address strings}
}
\value{
a character vector of standardized address strings
}
\description{
After cleaning and tagging addresses, specific tags are
concatenated to create a harmonized address suitable
for comparison and/or hashing.
Specific tags are separate by spaces; i.e., \verb{\{AddressNumber\} \{StreetName\} \{StreetNamePostType\} \{PlaceName\} \{StateName\} \{ZipCode\}}. All address tags are converted to lower case,
street name post types are expanded (e.g., "str" to "street" and "ave" to "avenue"),
and only the first five digits of the ZIP Code are used.
If any address tags are missing (except for \code{StreetNamePostType}), then a missing
standardized address will be returned.
}
\examples{
addr_standardize(
  x = c(
    "3333 Burnet Avenue Apt 2 Cincinnati OH 45220",
    "3333 bUrNeT Avenue Cincinnati OH 45220",
    "3333 Burnet Avenue Apt #2 Cincinnati OH 45220",
    "3333 Burnet Ave Cincinnati OH 45220",
    "3333 Burnet Av. Cincinnati OH 45220",
    "3333 Burnet Avenue Cincinnati 45220"
  )
)
}
