% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_street.R
\name{addr_street}
\alias{addr_street}
\title{Address street class}
\usage{
addr_street(
  predirectional = character(0),
  premodifier = character(0),
  pretype = character(0),
  name = character(0),
  posttype = character(0),
  postdirectional = character(0),
  map_posttype = TRUE,
  map_directional = TRUE,
  map_pretype = TRUE
)
}
\description{
Address street class
}
