% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/tag.R
\name{usaddress_tag}
\alias{usaddress_tag}
\alias{addr_tag}
\title{Return list of lists of address tags to R.}
\usage{
usaddress_tag(input)

addr_tag(x)
}
\arguments{
\item{input}{a character vector of addresses}

\item{x}{a character vector of addresses}
}
\value{
a list, the same length as x, of named character vectors of address component tags
}
\description{
Return list of lists of address tags to R.

Tag components of an address string
The address components are based upon the \href{http://www.urisa.org/advocacy/united-states-thoroughfare-landmark-and-postal-address-data-standard}{United States Thoroughfare, Landmark, and Postal Address Data Standard}
}
\examples{
addr_tag(c("224 Woolper Avenue Apt #2 Cincinnati OH 45220", "3333 Burnet Ave Cincinnati OH 45219"))
}
