% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{addr_tag}
\alias{addr_tag}
\title{Tag components of an address string
The address components are based upon the \href{http://www.urisa.org/advocacy/united-states-thoroughfare-landmark-and-postal-address-data-standard}{United States Thoroughfare, Landmark, and Postal Address Data Standard}}
\usage{
addr_tag(x)
}
\arguments{
\item{x}{a character vector of addresses}
}
\value{
a list, the same length as x, of lists of address component tags
}
\description{
Tag components of an address string
The address components are based upon the \href{http://www.urisa.org/advocacy/united-states-thoroughfare-landmark-and-postal-address-data-standard}{United States Thoroughfare, Landmark, and Postal Address Data Standard}
}
\examples{
addr_tag(c("224 Woolper Avenue Apt #2 Cincinnati OH 45220", "3333 Burnet Ave Cincinnati OH 45219"))
}
