% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr.R
\name{as_addr}
\alias{as_addr}
\title{coerce a character vector to an addr vector}
\usage{
as_addr(x)
}
\arguments{
\item{x}{a character vector or addr vector}
}
\value{
an addr vector
}
\description{
coerce a character vector to an addr vector
}
