% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_addr.R
\name{as_addr}
\alias{as_addr}
\title{Coerce to addr}
\usage{
as_addr(x, ...)
}
\arguments{
\item{x}{object to coerce to an addr}

\item{...}{additional arguments passed to methods}
}
\description{
\code{as_addr()} converts character vectors or data frames into \code{addr()} objects.
The structures for \code{addr()} and the \code{addr_} classes are derived as a subset
of the United States Thoroughfare, Landmark, and Postal Address Data Standard
that is relevant for residential thoroughfare addresses.
}
\section{Methods implemented for}{

\itemize{
\item \code{character}: will be cleaned, tagged, normalized; all \verb{map_*} arguments
are passed onto \code{addr_street()} or \code{addr_place()}
\item \code{data.frame}: must have columns named according to fields in
\code{addr_number()}, \code{addr_street()}, or \code{addr_place()}
\item \code{addr}: returned as-is
}
}

\examples{
as_addr(voter_addresses()[1:1000])

data.frame(
  number_digits = c("290", "200"),
  street_name = c("Burnet", "Main"),
  street_posttype = c("Ave", "St"),
  place_name = c("Cincinnati", "Cincinnati"),
  place_state = c("OH", "OH"),
  place_zipcode = c("45229", "45220"),
  stringsAsFactors = FALSE
)|>
  as_addr()
}
