% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_addr.R
\name{as_addr}
\alias{as_addr}
\title{Coerce to addr}
\usage{
as_addr(x, ...)
}
\arguments{
\item{x}{object to coerce to an addr vector}

\item{...}{additional arguments passed to methods}
}
\description{
\code{as_addr()} converts other objects into \code{addr()} vectors.
See \code{?addr} for more details on its structure.
}
\section{Methods implemented for}{

\itemize{
\item \code{character}: will be cleaned (if \code{clean = TRUE}) with \code{clean_address_text()}
and then tagged using \code{usaddress_tag()}; tags are normalized to abbreviations
by passing all \verb{map_*} arguments to \code{addr_street()} or \code{addr_place()};
ZIP codes parsed with more than five characters are truncated
with a warning; non-numeric characters in parsed
address number digits will be removed with a warning
\item \code{data.frame}: must have columns named according to fields in
\code{addr_number()}, \code{addr_street()}, or \code{addr_place()}; also passes
the \verb{map_*} arguments to \code{addr_street()} and \code{addr_place()}
\item \code{addr}: returned as-is
}
}

\examples{
as_addr(voter_addresses()[1:1000])

data.frame(
  number_digits = c("290", "200"),
  street_name = c("Burnet", "Main"),
  street_posttype = c("Ave", "St"),
  place_name = c("Cincinnati", "Cincinnati"),
  place_state = c("OH", "OH"),
  place_zipcode = c("45229", "45220"),
  stringsAsFactors = FALSE
)|>
  as_addr()
}
