% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cagis_addr}
\alias{cagis_addr}
\title{CAGIS Addresses}
\usage{
cagis_addr()
}
\value{
An example tibble created from the CAGIS addresses with a pre-calculated, unique \code{cagis_addr} vector column.
The \code{cagis_addr_data} column is a list of tibbles because one CAGIS address can correspond to multiple
parcel identifiers and address-level data (place, type, s2, etc.).
See \code{inst/make_cagis_addr.R} for source code to create data, including filtering criteria:
\itemize{
\item use only addresses that have \code{STATUS} of \code{ASSIGNED} or \code{USING} and are not orphaned (\code{ORPHANFLG == "N"})
\item omit addresses with \code{ADDRTYPE}s that are milemarkers (\code{MM}), parks (\code{PAR}), infrastructure projects (\code{PRJ}),
cell towers (\code{CTW}), vacant or commercial lots (\code{LOT}), and other miscellaneous non-residential addresses (\code{MIS}, \code{RR}, \code{TBA})
\item s2 cell is derived from LONGITUDE and LATITUDE fields in CAGIS address database
}
}
\description{
CAGIS Addresses
}
\examples{
cagis_addr()
}
