% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{expand_post_type}
\alias{expand_post_type}
\title{Expand street name post type}
\usage{
expand_post_type(x)
}
\arguments{
\item{x}{character vector of \code{StreetnamePostType} abbreviations}
}
\value{
a character vector of the same length containing the expanded street name post type
}
\description{
Abbreviations of street type (e.g., "Ave", "St") are converted to
expanded versions (e.g., "Avenue", "Street").
}
\examples{
expand_post_type(c("ave", "av", "Avenue", "tl"))
}
