% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Fuzzy match}
\usage{
fuzzy_match(x, y, osa_max_dist = 1, prefilter = c("none", "psk"))
}
\arguments{
\item{x}{character vector to match}

\item{y}{character vector to match to}

\item{osa_max_dist}{maximum OSA distance to consider a match;
\code{Inf} is a special case that avoids computing string distance by
returning all of \code{y} instead of just the best match or matches in \code{y}.}

\item{prefilter}{method used to prefilter \code{y} before computing
OSA distances; \code{"none"} does nothing, and \code{"psk"} removes values in \code{y}
that do not share a \code{phonetic_street_key()} with any value in \code{x}.}
}
\value{
a list of integer vectors representing the position of the best
matching string(s) in \code{y} for each string in \code{x}
}
\description{
Fuzzy match strings in \code{x} to strings in \code{y} using optimized string
alignment (OSA) distance and ignoring capitalization.
}
\details{
If multiple strings in \code{y} are tied for the minimum OSA distance from a
string in \code{x}, all of their indices are included in the return value.
}
\examples{
my_names <-
  c("Pinye", "Pine", "Oalck", "Sunset", "Riverbend", "Greenfild")
the_names <-
  c("Piney", "Pine", "Oak", "Cheshire", "Greenfield", "Maple", "Elm")
matches <- fuzzy_match(my_names, the_names, osa_max_dist = 1)
matches

lapply(matches, \(i) the_names[i])

x <- as_addr(voter_addresses()[1:100])@street@name
y <- unique(nad_example_data()$nad_addr@street@name)
system.time(fuzzy_match(x, y))
# larger vectors see a speedup when using
# phonetic_street_key as a prefilter
# but may miss potential matches that are within
# osa_max_dist of each other, but did not have
# identical phonetic codes (e.g., "woolper" and "woopler")
system.time(fuzzy_match(x, y, prefilter = "psk"))
}
