% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Fuzzy match}
\usage{
fuzzy_match(x, y, osa_max_dist = 1, prefilter = c("none", "psk"))
}
\arguments{
\item{x}{character vector to match}

\item{y}{character vector to match to}

\item{osa_max_dist}{maximum OSA distance to consider a match;
\code{Inf} is a special case that avoids computing string distance by
returning all of \code{y} instead of just the best match(es) in 'y`}

\item{prefilter}{method used to prefilter y before computing
osa distances for fuzzy matching to speed up calculations;
"none" does nothing, "psk" uses \code{phonetic_street_key()}}
}
\value{
a list of integer vectors representing the position of the best
matching string(s) in \code{y} for each string in \code{x}
}
\description{
Fuzzy match strings in x to y using optimized string alignment
(ignoring capitalization).
}
\details{
If multiple strings in \code{y} are tied for the minimum osa distances with a
string in \code{x}, then all of their indices are included in the return value.
}
\examples{
my_names <-
  c("Pinye", "Pine", "Oalck", "Sunset", "Riverbend", "Greenfild")
the_names <-
  c("Piney", "Pine", "Oak", "Cheshire", "Greenfield", "Maple", "Elm")
matches <- fuzzy_match(my_names, the_names, osa_max_dist = 1)
matches

lapply(matches, \(i) the_names[i])

# larger vectors see a speedup when using
# phonetic_street_key as a prefilter
x <- as_addr(voter_addresses()[1:100])@street@name
y <- nad_example_data()$nad_addr@street@name
system.time(fuzzy_match(x, y))
system.time(fuzzy_match(x, y, prefilter = "psk"))
}
