% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzymatch.R
\name{fuzzy_match}
\alias{fuzzy_match}
\alias{fuzzy_match_addr_field}
\title{fuzzy match strings in x to y using optimized string alignment (ignoring capitalization)}
\usage{
fuzzy_match(x, y, osa_max_dist = 1, ties = c("first", "random", "all"))

fuzzy_match_addr_field(
  x_addr,
  y_addr,
  addr_field,
  osa_max_dist = 0,
  ties = "all"
)
}
\arguments{
\item{x}{character vector to match}

\item{y}{character vector to match to}

\item{osa_max_dist}{maximum osa distance to consider a match}

\item{ties}{if multiple strings in \code{y} are tied for the minimum osa distances with a string in \code{x},
then specify "first" or "random" as a tiebreaker}

\item{x_addr}{addr vector to match}

\item{y_addr}{addr vector to match to}

\item{addr_field}{character name of \code{addr()} field to match on}
}
\value{
an integer vector representing the position of the best matching string in \code{y} for each string in \code{x};
when \code{ties} is "all", a list of integer vectors is returned instead
}
\description{
fuzzy match strings in x to y using optimized string alignment (ignoring capitalization)
}
\details{
\code{fuzzy_match_addr_field} is a helper to match addr vectors using fuzzy_match on a specific field
}
