% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\alias{geocode_zip}
\title{Geocode addr vectors with census.gov TIGER address features}
\usage{
geocode(x, offset = 0L)

geocode_zip(x, offset = 0L)
}
\arguments{
\item{x}{an addr vector (\code{?as_addr})}

\item{offset}{number of meters to offset geocode from street line}
}
\value{
a tbl with columns: addr (\code{x}, addr vector),
matched zipcode (character vector, matched street (addr_street vector),
s2_geography (s2_geography point vector), and s2_cell (s2_cell vector)
}
\description{
\code{geocode_tiger()} geocodes addr vectors using census.gov TIGER address
features (see \code{?taf}) by:
\enumerate{
\item searching for a matching street (see \code{?match_addr_street}),
within the same zip code, also searching similar zip codes (see
\code{?zip_variant}) for a matching street if necessary
\item using the address number to select the best address feature range and
side of the street (even/odd), tiebreaking on smallest width and spread)
\item linearly interpolate a geographic point along the best range line based
on the actual and potential range of address numbers
}

Only matched input addr will return non-missing matched zipcode/street
values. Missing or unmatched zip codes return missing matched zipcode/street,
geography, and s2cell values, like any other non-match. If all ranges on the
matched zipcode/street exclude the address number, then \emph{only} the
geography/s2cell values will return NA.
}
\details{
\code{geocode_zip()} is the workhorse function and operates on addr vectors
with the same ZIP code; use \code{geocode()} to geocode an addr vector
with different ZIP codes by grouping them by ZIP code and processing
serially.
At a lower level, grouping addr vectors by ZIP code and applying
\code{geocode_zip()} facilitates more control (e.g., parallel processing)
}
\examples{

x <- as_addr(voter_addresses()[1:100])
gcd <- geocode(x)

leaflet::leaflet(wk::wk_coords(gcd$matched_geography)) |>
  leaflet::addTiles() |>
  leaflet::addCircleMarkers(lng = ~x, lat = ~y, label = ~feature_id)
}
