% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_part_match.R
\name{match_addr_number}
\alias{match_addr_number}
\title{Match addr_number vectors}
\usage{
match_addr_number(x, y, number_fuzzy_dist = 1L)
}
\arguments{
\item{x, y}{addr_number vectors to match}

\item{number_fuzzy_dist}{integer; maximum optimized string alignment distance
between \verb{@number} of x and y to consider a possible match}
}
\value{
an addr_number vector, the same length as x, that is the
best match in y for each addr_number code in x; if no best match
is found a missing value is returned (\code{addr_number()})
}
\description{
A single addr_number in y is chosen for each addr_number in x.
If exact matches (using \code{as.character}) are not found,
possible matches (within \code{number_fuzzy_distance}) are searched for in y.
If multiple matches are present in y, the best one is selected
based on the lowest absolute numeric difference with the @digits in x;
ties are broken by optimized string alignment (OSA) distances
and then preferring the lowest value sorted in Lexicographic
order with digits preceding alphabetic characters.

addr_number objects with missing @digits or empty strings
for all of @prefix, @digits, @suffix are not matched and
returned as missing instead.
}
\examples{
 x <- addr_number(
   prefix = "",
   digits = as.character(c(1, 10, 228, 11, 22, 22, 22, 10, 99897, NA)),
   suffix = ""
 )

y <- addr_number(
  prefix = "",
  digits = as.character(c(12, 11, 10, 22)),
  suffix = ""
)

match_addr_number(x, y)

match_addr_number(x, y, number_fuzzy_dist = 0L)
}
