% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_part_match.R
\name{match_zipcodes}
\alias{match_zipcodes}
\title{Match ZIP codes}
\usage{
match_zipcodes(x, y, zip_variants = TRUE)
}
\arguments{
\item{x, y}{character vectors of ZIP codes to match}

\item{zip_variants}{logical; fuzzy match to common variants of
\code{x} in \code{y}? (for example, changing the fourth or fifth digit)}
}
\value{
A character vector, the same length as \code{x}, containing the selected
match in \code{y} for each ZIP code in \code{x}.
}
\description{
A single ZIP code in \code{y} is chosen for each ZIP code in \code{x}.
By default, if exact matches are not found, common variants of ZIP codes
in \code{x} are searched for in \code{y} (?\code{zipcode_variant})
If multiple variants are present in \code{y}, the selected match has the lowest
absolute numeric difference from the ZIP code in \code{x}; ties are broken by
OSA string distance and then by the
minimum number.
}
\examples{
match_zipcodes(
  c("45222", "45219", "45219", "45220", "45220", "", NA),
  c("42522", "45200", "45219", "45221", "45223", "45321", "")
)

match_zipcodes(
  c("45222", "45219", "45219", "45220", "45220", "", NA),
  c("42522", "45200", "45219", "45221", "45223", "45321", ""),
  zip_variants = FALSE
)
}
