% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nad.R
\name{nad}
\alias{nad}
\alias{nad_read}
\alias{nad_download}
\title{Read National Address Database (NAD) tables into R}
\usage{
nad(county, state, refresh = c("yes", "no", "force"))

nad_read(county, state)

nad_download()
}
\arguments{
\item{county}{character, length one; name of county}

\item{state}{character, length one; name of state}

\item{refresh}{character, length one; choose how to refresh NAD
data binaries cached on disk if not already present; "yes" will
create data binary if not already present, "no" will
error if data binary is not already present, "force" will
create the data binary and overwrite any existing data binary}
}
\description{
The U.S. Department of Transportation partners with address programs from
state, local, and tribal governments to compile their authoritative data
into a database. Find more information here:
\url{https://www.transportation.gov/gis/national-address-database}

\code{nad_read()} reads data from the NAD geodatabase by county and state,
downloading it first to the R user's data directory for the addr
package if not already downloaded with \code{nad_download()},
and readies it for R.
The NAD geodatabase has a very large size on disk (~10 GB).

Data binaries are the cached outputs of \code{nad_read()} for each
County/State and are created on first run with \code{nad()}.
Download data binaries to the R_user_dir cache folder or point R
(\code{?R_user_dir}) to these files on disk to read NAD tables without
having to download the nationwide NAD geodatabase.
(\code{list.files(tools::R_user_dir("addr", "cache"), recursive = TRUE)})
}
\details{
The NAD is downloaded as the latest release from the transportation.gov
data portal:
\url{https://data.transportation.gov/dataset/National-Address-Database-NAD-File-Geodatabase/yw36-suxr/about_data}
For the original schema, see
\url{https://www.transportation.gov/sites/dot.gov/files/2023-07/NAD_Schema_202304.pdf}

The NAD does not distinguish between empty and missing address components.
When reading into R, all missing address components are replaced with an
empty string (\code{""}) \emph{except} for address number (digits), street name,
and ZIP code.
}
\examples{
# read data from NAD, caching output on first run
\dontrun{
  nad("Hamilton", "OH")
}

# example data preloaded for Hamilton County, OH
# works without downloading NAD gdb first
nad("Hamilton", "OH", refresh = "no")
}
