% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{nad_example_data}
\alias{nad_example_data}
\title{Example National Address Database addresses}
\usage{
nad_example_data()
}
\value{
a tibble with 349,407 rows and 7 columns
}
\description{
An example of the data returned using \code{nad_read()} for
Hamilton County, Ohio.  See \code{?nad_read} for more information about
the National Address Database.
}
\examples{
nad_example_data()
}
