% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_addr_feat.R
\name{taf}
\alias{taf}
\alias{taf_install}
\title{TIGER Address Features dataset}
\usage{
taf(year = as.character(2025:2011), version = "v1")

taf_install(county_fips, year, version = "v1")
}
\arguments{
\item{year}{integer, length one; vintage of TIGER addrfeat files}

\item{version}{character, length one; major version of the package
and taf dataset schema}

\item{county_fips}{character, length 1; county fips code}
}
\value{
(from the arrow package), A Dataset R6 object. Use \code{dplyr} methods
on it to query the data, or call \verb{$NewScan()} to construc a query directly.
}
\description{
taf() uses the arrow package to open the hive partitioned, parquet dataset
of TIGER address features in the R user directory for the addr package
}
\examples{
taf()
taf_install("39061", "2025")
}
