% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.R
\name{taf}
\alias{taf}
\alias{taf_install}
\title{TIGER Address Features dataset}
\usage{
taf(year = as.character(2025:2011), version = "v1")

taf_install(
  county,
  year = as.character(2025:2011),
  version = "v1",
  overwrite = FALSE,
  redownload = FALSE
)
}
\arguments{
\item{year}{integer, length one; vintage of TIGER addrfeat files}

\item{version}{character, length one; major version of the package
and taf dataset schema}

\item{county}{character, length 1; county fips code}

\item{overwrite}{logical, length 1; overwrite an existing county install?}

\item{redownload}{logical, length 1; re-download cached TIGER zip files?}
}
\value{
a Dataset R6 object (see \code{?arrow::open_dataset}); use \code{dplyr}
verbs to query the data and get results, see examples
}
\description{
taf() uses the arrow package to open the
hive partitioned, parquet dataset of TIGER address features in the R user
directory for the addr package.
The FileSystemDataset in Arrow is database-like backend for
larger-than-memory datasets and provides functionality
to use dplyr syntax for data manipulation;
see \url{https://arrow.apache.org/docs/r/articles/data_wrangling.html} for
more details on arrow and dplyr.

\code{install_taf()} will download and link TIGER address features and
feature names for a specific year and county, installing the resulting
file in the R user's data directory for the addr package.
If an address feature does not have a corresponding LINEARID with a
feature name, then the street tags are parsed from the full name, in
which case the column, \code{street_tag_parsed} will be TRUE.
}
\examples{
\dontrun{
taf()

# use dplyr verbs to query
library(dplyr, warn.conflicts = FALSE)

# find top ten most frequent street name-posttype combinations
taf() |>
  group_by(street_name, street_posttype) |>
  summarize(
    n_zips = n_distinct(ZIP),
    n_ranges = n(),
    .groups = "drop"
  ) |>
  arrange(desc(n_zips), desc(n_ranges)) |>
  collect() |>
  slice(1:10)
}
\dontrun{
  taf_install("39061", "2025")
}
}
