% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_addr_feat.R
\name{taf}
\alias{taf}
\alias{taf_install}
\title{TIGER Address Features dataset}
\usage{
taf(year = as.character(2025:2011), version = "v1")

taf_install(county_fips, year, version = "v1")
}
\arguments{
\item{year}{integer, length one; vintage of TIGER addrfeat files}

\item{version}{character, length one; major version of the package
and taf dataset schema}

\item{county_fips}{character, length 1; county fips code}
}
\value{
a Dataset R6 object (see \code{?arrow::open_dataset}); use \code{dplyr}
verbs to query the data and get results, see examples
}
\description{
taf() uses the arrow package to open the
hive partitioned, parquet dataset of TIGER address features in the R user
directory for the addr package.
The FileSystemDataset in Arrow is database-like backend for
larger-than-memory datasets and provides functionality
to use dplyr syntax for data manipulation;
see \url{https://arrow.apache.org/docs/r/articles/data_wrangling.html} for
more details on arrow and dplyr
}
\examples{
taf()

# use dplyr verbs to query and
# find top ten most frequent street name-posttype combinations
library(dplyr, warn.conflicts = FALSE)
taf() |>
  group_by(street_name, street_posttype) |>
  summarize(
    n_zips = n_distinct(ZIP),
    n_ranges = n(),
    .groups = "drop"
  ) |>
  arrange(desc(n_zips), desc(n_ranges)) |>
  collect() |>
  slice(1:10)
taf_install("39061", "2025")
}
