% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.R
\name{taf_dataset}
\alias{taf_dataset}
\title{Open installed TIGER Address Features as an Arrow dataset}
\usage{
taf_dataset(year = as.character(2025:2011), version = "v2")
}
\arguments{
\item{year}{character, length one; vintage of TIGER addrfeat (address feature)
files}

\item{version}{character, length one; major version of the package
and taf dataset schema}
}
\value{
a Dataset R6 object (see \code{?arrow::open_dataset}); use \code{dplyr}
verbs to query the data and get results, see examples
}
\description{
\code{taf_dataset()} uses the arrow package to open the Hive-partitioned Parquet
dataset of TIGER address features in the addr user data directory.
Arrow \code{FileSystemDataset} objects are database-like backends for
larger-than-memory datasets and support dplyr syntax for data manipulation;
see \url{https://arrow.apache.org/docs/r/articles/data_wrangling.html}.
Other TAF helpers such as \code{taf_catalog()}, \code{taf_install()}, and \code{taf()}
use nanoparquet directly for flat parquet file reads and writes. Arrow is
only required for the advanced dataset interface returned by
\code{taf_dataset()}.
}
\examples{
\dontrun{
  Sys.setenv("R_USER_DATA_DIR" = tempfile())
  taf_install("39061", "2025")

  if (requireNamespace("arrow", quietly = TRUE) &&
    requireNamespace("dplyr", quietly = TRUE)) {
    taf_dataset()

    # find top ten most frequent street name-posttype combinations
    taf_dataset() |>
      dplyr::group_by(street_name, street_posttype) |>
      dplyr::summarize(
        n_zips = dplyr::n_distinct(ZIP),
        n_ranges = dplyr::n(),
        .groups = "drop"
      ) |>
      dplyr::arrange(dplyr::desc(n_zips), dplyr::desc(n_ranges)) |>
      dplyr::collect() |>
      dplyr::slice(1:10)
  }
}
}
