% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.R
\name{taf_zip}
\alias{taf_zip}
\title{Read taf() data for ZIP codes across all installed counties}
\usage{
taf_zip(x, map = TRUE, year = as.character(2025:2011), version = "v1")
}
\arguments{
\item{x}{character vector of five-digit ZIP codes}

\item{map}{logical, length 1; map street tags read from taf() data
(type, directional, ordinal) when converting to \code{addr_street()} vector?}

\item{year}{character, length 1; vintage of TIGER addrfeat (address feature)
files}

\item{version}{character, length 1; major version of the package
and taf dataset schema}
}
\value{
a tibble with \code{LINEARID}, \code{FULLNAME}, \code{side}, \code{ZIP},
\code{FROMHN}, \code{TOHN}, \code{PARITY}, \code{OFFSET}, \code{s2_geography}, \code{addr_street},
\code{county_fips}, and \code{street_tag_parsed} columns
}
\description{
\code{taf_zip()} reads and transforms \code{taf()} data for a subset of ZIP codes.
It reconstructs the \code{county_fips}, \code{s2_geography}, and \code{addr_street}
vectors in the returned data frame.
}
\examples{
Sys.setenv("R_USER_DATA_DIR" = tempfile())
taf_install("39061", "2025")
taf_zip(c("45249", "45230", "45220"))
}
