% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taf.R
\name{taf_zip}
\alias{taf_zip}
\title{read taf() for a ZIP code across all installed counties}
\usage{
taf_zip(x, map = TRUE, year = as.character(2025:2011), version = "v1")
}
\arguments{
\item{map}{logical, length 1; map street tags read from taf() data
(type, directional, ordinal) when converting to \code{addr_street()} vector?}

\item{year}{character, length 1; vintage of TIGER addrfeat files}

\item{version}{character, length 1; major version of the package
and taf dataset schema}
}
\value{
a tibble with \code{LINEARID}, \code{FULLNAME}, \code{side}, \code{ZIP},
\code{FROMHN}, \code{TOHN}, \code{PARITY}, \code{OFFSET}, \code{s2_geography}, \code{addr_street},
\code{county_fips}, and \code{street_tag_parsed} columns
}
\description{
taf_zip() is a helper that transforms data read in from taf()
for a subset of zip codes to use with the addr package.
It reconstructs the \code{county_fips}, \code{s2_geography}, and \code{addr_street},
vectors in the returned data frame to use with the addr package.
}
\examples{
\dontrun{
  taf_zip(c("45249", "45230", "45220"))
}
}
