% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_usaddress.R
\name{tag_usaddress}
\alias{tag_usaddress}
\title{Tag US addresses}
\usage{
tag_usaddress(x = NA_character_, clean = TRUE)
}
\arguments{
\item{x}{character string of addresses}

\item{clean}{logical; clean address text with clean_address_text()
before tagging?}
}
\value{
a list of vectors of named address tags
}
\description{
Addresses are tagged using the usaddress conditional random field
in a \href{https://github.com/boydjohnson/usaddress-rs}{rust port} of
\href{https://github.com/datamade/usaddress}{usaddress}.
Possible address labels include:
\itemize{
\item \code{AddressNumberPrefix}
\item \code{AddressNumberSuffix}
\item \code{AddressNumber}
\item \code{BuildingName}
\item \code{CornerOf}
\item \code{IntersectionSeparator}
\item \code{LandmarkName}
\item \code{NotAddress}
\item \code{OccupancyIdentifier}
\item \code{OccupancyType}
\item \code{PlaceName}
\item \code{Recipient}
\item \code{StateName}
\item \code{StreetNamePostDirectional}
\item \code{StreetNamePostType}
\item \code{StreetNamePreDirectional}
\item \code{StreetNamePreModifier}
\item \code{StreetNamePreType}
\item \code{StreetName}
\item \code{SubaddressIdentifier}
\item \code{SubaddressType}
\item \code{USPSBoxGroupID}
\item \code{USPSBoxGroupType}
\item \code{USPSBoxID}
\item \code{USPSBoxType}
\item \code{ZipCode}
}

Find more information about the definitions at
\url{https://www.fgdc.gov/standards/projects/address-data}
}
\examples{
tag_usaddress(
  c("290 Ludlow Avenue Apt 2 Cincinnati OH 45220",
  "3333 Burnet Ave Cincinnati Ohio 45219",
  "120 North Main Street, Greenville, SC 29601",
  "200 Southwest North Street, Topeka, KS 66603",
  "215 Highway 88 Road, Jackson, CA 95642"
  )
)

# edge cases!
tag_usaddress(
  c(
    "1600 Pennsylvania Avenue NW, Washington, DC 20500", # post-directional quadrant
    "1 Infinite Loop, Cupertino, CA 95014", # corporate campus street name
    "210 East 400 South, Salt Lake City, UT 84111", # grid addressing (Utah)
    "N6W23001 Bluemound Road, Wauwatosa, WI 53226", # address number prefix grid (Wisconsin)
    "350 Fifth Avenue, New York, NY 10118", # ordinal street name
    "4059 Mt Lee Drive, Hollywood, CA 90068", # abbreviated street element
    "233 South Wacker Drive, Chicago, IL 60606", # pre-directional
    "700 Exposition Park Drive, Los Angeles, CA 90037", # multi-word street name
    "2 South Biscayne Boulevard, Miami, FL 33131" # directional + boulevard
  )
)
}
