% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger.R
\name{tiger_addr_feat}
\alias{tiger_addr_feat}
\title{Get s2_geography for tiger street ranges}
\usage{
tiger_addr_feat(county, year, redownload = FALSE)
}
\arguments{
\item{county}{character string of county identifier}

\item{year}{character year of tigris product}

\item{redownload}{logical, length 1; re-download the cached TIGER zip file?}
}
\value{
a tibble with \code{LINEARID}, \code{FULLNAME}, \code{side}, \code{ZIP},
\code{FROMHN}, \code{TOHN}, \code{PARITY}, \code{OFFSET}, and \code{s2_geography} columns
}
\description{
TIGER address features (i.e. street address ranges) are read from compressed
addrfeat shapefiles for each county and Census vintage.
If not already present, compressed addrfeat shapefiles are downloaded
from the census.gov FTP site to the R user's data directory for
the addr package.

When reading into R, the data is converted to a "long" format by street
side L/R to be used within \code{taf_install()} to create \code{taf()}.
}
\examples{
tiger_addr_feat("39061", "2025")
}
