% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger.R
\name{tiger_addr_feat}
\alias{tiger_addr_feat}
\title{Get s2_geography for tiger street ranges}
\usage{
tiger_addr_feat(county, year)
}
\arguments{
\item{county}{character string of county identifier}

\item{year}{year of tigris product}
}
\value{
a list of tibbles, one for each street name, with \code{TLID}, \code{s2_geography}, \code{from}, and \code{to} columns
}
\description{
TIGER address features (i.e. street address ranges) are downloaded
}
\keyword{internal}
