% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_geography.R
\name{tiger_download}
\alias{tiger_download}
\title{download tiger files
files are saved to the R user cache directory;
see \code{?tools::R_user_dir()} to change this}
\usage{
tiger_download(x)
}
\arguments{
\item{x}{filename relative to https://www2.census.gove/geo/tiger/}
}
\description{
download tiger files
files are saved to the R user cache directory;
see \code{?tools::R_user_dir()} to change this
}
\examples{
Sys.setenv("R_USER_CACHE_DIR" = tempdir())
tiger_download("TIGER2022/ADDRFEAT/tl_2022_39061_addrfeat.zip")
}
