% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger.R
\name{tiger_feat_names}
\alias{tiger_feat_names}
\title{Get names for tiger street ranges}
\usage{
tiger_feat_names(county, year, redownload = FALSE)
}
\arguments{
\item{county}{character string of county FIPS identifier}

\item{year}{character year of the Census TIGER/Line product}

\item{redownload}{logical, length 1; replace the durable managed local copy
of the TIGER ZIP file?}
}
\value{
a tibble with unique \code{LINEARID} and \code{addr} columns
}
\description{
TIGER primary feature names are read from compressed feature-name databases
for each county and Census vintage.
If not already present, compressed addrfeat (address feature) shapefiles are
downloaded from the Census TIGER HTTPS endpoint by default as durable
managed local copies in addr's \code{stow/tiger_feat_names} directory. Files from
the former unmanaged TIGER layout are not searched.

When reading into R, the data is filtered to addressable MTFCCs
(S1100, S1200, S1400, S1640) that have a name.
}
\examples{
\dontrun{
  tiger_feat_names("39061", "2025")
}
}
