% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger.R
\name{tiger_feat_names}
\alias{tiger_feat_names}
\title{Get names for tiger street ranges}
\usage{
tiger_feat_names(county, year, redownload = FALSE)
}
\arguments{
\item{county}{character string of county identifier}

\item{year}{character year of tigris product}

\item{redownload}{logical, length 1; re-download the cached TIGER zip file?}
}
\value{
a tibble with unique \code{LINEARID} and \code{addr} columns
}
\description{
TIGER primary feature names are read from compressed featnames databases for
each county and census vintage.
If not already present, compressed addrfeat shapefiles are downloaded
from the census.gov FTP site to the R user's data directory for
the addr package.

When reading into R, the data is filtered to addressable MTFCCs
(S1100, S1200, S1400, S1640) that have a name.
}
\examples{
tiger_feat_names("39061", "2025")
}
