% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{usaddress_tag}
\alias{usaddress_tag}
\title{Return list of lists of address tags to R.}
\usage{
usaddress_tag(input)
}
\arguments{
\item{input}{character string of addresses}
}
\description{
Return list of lists of address tags to R.
}
