% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{voter_addresses}
\alias{voter_addresses}
\title{Example real-world addresses}
\usage{
voter_addresses()
}
\value{
a character vector
}
\description{
The voter_addresses data was generated as an example character vector of real-world addresses.
These addresses were downloaded from the Hamilton County, Ohio voter registration database on 2024-09-12.
See \code{inst/make_example_addresses.R} for more details.
\code{AddressPreDirectional}, \code{AddressNumber}, \code{AddressStreet}, \code{AddressSuffix}, \code{CityName}, "OH", and \code{AddressZip}
are pasted together to create 242,133 unique addresses of registered voters in Hamilton County, OH.
}
\examples{
voter_addresses()
}
