% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{voter_addresses}
\alias{voter_addresses}
\title{Example addresses}
\usage{
voter_addresses()
}
\value{
a character vector
}
\description{
\code{voter_addresses()} returns an example character vector of real-world
addresses downloaded from the Hamilton County, Ohio voter registration
database on 2024-09-12. \code{AddressPreDirectional}, \code{AddressNumber},
\code{AddressStreet}, \code{AddressSuffix}, \code{CityName}, \code{"OH"}, and \code{AddressZip}
were pasted together to create 242,133 unique registered-voter addresses.
}
\examples{
voter_addresses() |>
  head()
}
