% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_usps_zip_code_crosswalk.R
\name{zip_fips_lookup}
\alias{zip_fips_lookup}
\title{Translate ZIP codes and county FIPS identifiers}
\usage{
zip_fips_lookup(zip, fips = NULL)
}
\arguments{
\item{zip}{character, length one; a 5-digit ZIP code}

\item{fips}{character, length one; a 5-digit FIPS county identifier}
}
\value{
if fips is NULL, a character vector of matched 5-digit FIPS
county identifiers, ordered by percentages of residential addresses;
if fips is not NULL, an addr_place vector of length one
}
\description{
\code{zip_fips_lookup()} uses a package-internal reference derived from the
2025 Q4 HUD USPS ZIP Code Crosswalk Files to translate between
ZIP codes and county FIPS identifiers.

Only one of zip and fips should be provided and the other will be
returned.  ZIP codes and counties are not mutually exclusive;
output could be longer than one if multiple matches are present.
}
\examples{
zip_fips_lookup(zip = "45220")
zip_fips_lookup(zip = "45249")
zip_fips_lookup(zip = "45249", fips = "39061")
}
